Import version 4.4.4.2

+++ classes/evalcomix_assessments.php
@@ -19,26 +19,38 @@ require_once('evalcomix_object.php');
 require_once('evalcomix_modes.php');
 
 /**
+ * EvalCOMIX assessments
  * @package    block_evalcomix
  * @copyright  2010 onwards EVALfor Research Group {@link http://evalfor.net/}
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  * @author     Daniel Cabeza Sánchez <daniel.cabeza@uca.es>, Juan Antonio Caballero Hernández <juanantonio.caballero@uca.es>
  */
 
+/**
+ * EvalCOMIX assessments
+ * @package    block_evalcomix
+ * @copyright  2010 onwards EVALfor Research Group {@link http://evalfor.net/}
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ * @author     Daniel Cabeza Sánchez <daniel.cabeza@uca.es>, Juan Antonio Caballero Hernández <juanantonio.caballero@uca.es>
+ */
 class block_evalcomix_assessments extends block_evalcomix_object {
+    /**
+     * Table name
+     * @var string $table
+     */
     public $table = 'block_evalcomix_assessments';
 
     /**
      * Array of required table fields, must start with 'id'.
      * @var array $requiredfields
      */
-    public $requiredfields = array('id', 'taskid', 'assessorid', 'studentid', 'grade', 'timemodified');
+    public $requiredfields = ['id', 'taskid', 'assessorid', 'studentid', 'grade', 'timemodified'];
 
     /**
      * Array of optional table fields.
      * @var array $requiredfields
      */
-    public $optionalfields = array();
+    public $optionalfields = [];
 
     /**
      * course_module ID associated
@@ -87,7 +99,7 @@ class block_evalcomix_assessments extends block_evalcomix_object {
         $this->timemodified = 0;
         // Por si queremos crear una instancia vacía (para usar evalcomix_object::fetch_all_helper es necesario).
         if (is_numeric($taskid) && !is_float($taskid) && (int)$taskid > 0) {
-            $task = $DB->get_record('block_evalcomix_tasks', array('id' => $taskid), '*', MUST_EXIST);
+            $task = $DB->get_record('block_evalcomix_tasks', ['id' => $taskid], '*', MUST_EXIST);
             $this->taskid = $task->id;
         } else {
             $this->taskid = 0;
@@ -95,7 +107,7 @@ class block_evalcomix_assessments extends block_evalcomix_object {
 
         // Por si queremos crear una instancia vacía (para usar evalcomix_object::fetch_all_helper es necesario).
         if (is_numeric($assessorid) && !is_float($assessorid) && $assessorid > '0') {
-            $assessor = $DB->get_record('user', array('id' => $assessorid), '*', MUST_EXIST);
+            $assessor = $DB->get_record('user', ['id' => $assessorid], '*', MUST_EXIST);
             $this->assessorid = $assessor->id;
         } else {
             $this->assessorid = 0;
@@ -103,7 +115,7 @@ class block_evalcomix_assessments extends block_evalcomix_object {
 
         // Por si queremos crear una instancia vacía (para usar evalcomix_object::fetch_all_helper es necesario).
         if (is_numeric($studentid) && !is_float($studentid) && $studentid > '0') {
-            $student = $DB->get_record('user', array('id' => $studentid), '*', MUST_EXIST);
+            $student = $DB->get_record('user', ['id' => $studentid], '*', MUST_EXIST);
             $this->studentid = $student->id;
         } else {
             $this->studentid = 0;
@@ -112,7 +124,6 @@ class block_evalcomix_assessments extends block_evalcomix_object {
 
     /**
      * Finds and returns all evalcomix_assessments instances.
-     * @static abstract
      * @param array $params
      * @return array array of evalcomix_assessments instances or false if none found.
      */
@@ -121,42 +132,43 @@ class block_evalcomix_assessments extends block_evalcomix_object {
     }
 
     /**
+     * get_assessments_by_modality
      * @param int $taskid
      * @param int $userid
      * @return object with evalcomix_assessment objects by modality and their weighings
      */
     public static function get_assessments_by_modality($taskid, $userid) {
         global $CFG, $DB;
-        require_once($CFG->dirroot .'/blocks/evalcomix/classes/evalcomix_modes.php');
-        require_once($CFG->dirroot .'/blocks/evalcomix/classes/evalcomix_tasks.php');
-        $assessments = $DB->get_records('block_evalcomix_assessments', array('studentid' => $userid, 'taskid' => $taskid));
+        require_once($CFG->dirroot . '/blocks/evalcomix/classes/evalcomix_modes.php');
+        require_once($CFG->dirroot . '/blocks/evalcomix/classes/evalcomix_tasks.php');
+        $assessments = $DB->get_records('block_evalcomix_assessments', ['studentid' => $userid, 'taskid' => $taskid]);
         if ($assessments) {
-            if (!$task = $DB->get_record('block_evalcomix_tasks', array('id' => $taskid))) {
+            if (!$task = $DB->get_record('block_evalcomix_tasks', ['id' => $taskid])) {
                 return false;
             }
-            $cm = $DB->get_record('course_modules', array('id' => $task->instanceid));
+            $cm = $DB->get_record('course_modules', ['id' => $task->instanceid]);
 
             $context = context_course::instance($cm->course);
             $selfassessment = null;
-            $teacherassessment = array();
-            $peerassessments = array();
+            $teacherassessment = [];
+            $peerassessments = [];
             $weighingteacher = null;
             $weighingself = null;
             $weighingpeer = null;
 
             foreach ($assessments as $assessment) {
                 if (has_capability('moodle/grade:viewhidden', $context, $assessment->assessorid)) {
-                    if ($modality = $DB->get_record('block_evalcomix_modes', array('taskid' => $taskid, 'modality' => 'teacher'))) {
+                    if ($modality = $DB->get_record('block_evalcomix_modes', ['taskid' => $taskid, 'modality' => 'teacher'])) {
                         $weighingteacher = $modality->weighing;
                         array_push($teacherassessment, $assessment);
                     }
                 } else if ($assessment->assessorid == $userid) {
-                    if ($modality = $DB->get_record('block_evalcomix_modes', array('taskid' => $taskid, 'modality' => 'self'))) {
+                    if ($modality = $DB->get_record('block_evalcomix_modes', ['taskid' => $taskid, 'modality' => 'self'])) {
                         $weighingself = $modality->weighing;
                         $selfassessment = $assessment;
                     }
                 } else {
-                    if ($modality = $DB->get_record('block_evalcomix_modes', array('taskid' => $taskid, 'modality' => 'peer'))) {
+                    if ($modality = $DB->get_record('block_evalcomix_modes', ['taskid' => $taskid, 'modality' => 'peer'])) {
                         $weighingpeer = $modality->weighing;
                         array_push($peerassessments, $assessment);
                     }
@@ -177,12 +189,13 @@ class block_evalcomix_assessments extends block_evalcomix_object {
     }
 
     /**
+     * calculate_gradearray
      * @param array $assessments array of evalcomix_assessments objects
      * @return double|false result of current implementation of icalculator interface of false if $assessment is empty
      */
     public static function calculate_gradearray($assessments) {
         if (!empty($assessments)) {
-            $grades = array();
+            $grades = [];
             foreach ($assessments as $assessment) {
                 array_push($grades, $assessment->grade);
             }
@@ -193,12 +206,13 @@ class block_evalcomix_assessments extends block_evalcomix_object {
     }
 
     /**
+     * calculate_grades
      * @param array $assessments array of evalcomix_assessments objects
      * @return array|false result of current implementation of icalculator interface of false if $assessment is empty
      */
     public static function calculate_grades($assessments) {
         if (!empty($assessments)) {
-            $grades = array();
+            $grades = [];
             foreach ($assessments as $assessment) {
                 array_push($grades, $assessment->grade);
             }
@@ -208,12 +222,13 @@ class block_evalcomix_assessments extends block_evalcomix_object {
     }
 
     /**
+     * get_students_assessed
      * @param int $taskid
      * @return array of IDs of student assessed in $taskid
      */
     public static function get_students_assessed($taskid) {
-        $assessments = self::fetch_all(array('taskid' => $taskid));
-        $students = array();
+        $assessments = self::fetch_all(['taskid' => $taskid]);
+        $students = [];
         if ($assessments) {
             foreach ($assessments as $assessment) {
                 array_push($students, $assessment->studentid);
@@ -224,54 +239,65 @@ class block_evalcomix_assessments extends block_evalcomix_object {
         return false;
     }
 
+    /**
+     * delete_assessment_by_modeid
+     */
     public static function delete_assessment_by_modeid($modeid) {
         global $CFG, $DB;
         $result = true;
-        if ($DB->get_records('block_evalcomix_modes', array('id' => $modeid))) {
-            if ($assessments = $DB->get_records('block_evalcomix_assessments', array('modeid' => $modeid))) {
-                $cms = array();
+        if ($DB->get_records('block_evalcomix_modes', ['id' => $modeid])) {
+            if ($assessments = $DB->get_records('block_evalcomix_assessments', ['modeid' => $modeid])) {
+                $cms = [];
                 foreach ($assessments as $key => $assessment) {
                     $taskid = $assessment->taskid;
                     if (empty($cms[$taskid])) {
-                        if ($task = $DB->get_record('block_evalcomix_tasks', array('id' => $taskid))) {
-                            if ($cm = $DB->get_record('course_modules', array('id' => $task->instanceid))) {
+                        if ($task = $DB->get_record('block_evalcomix_tasks', ['id' => $taskid])) {
+                            if ($cm = $DB->get_record('course_modules', ['id' => $task->instanceid])) {
                                 $cms[$taskid] = $cm;
                             }
                         }
                     }
                 }
-                $result = self::delete_assessment(array('where' => array('modeid' => $modeid), 'cmid' => $cms[$taskid]->id,
-                    'courseid' => $cms[$taskid]->course));
+                $result = self::delete_assessment(['where' => ['modeid' => $modeid], 'cmid' => $cms[$taskid]->id,
+                    'courseid' => $cms[$taskid]->course]);
             }
         }
         return $result;
     }
 
+    /**
+     * delete_assessment
+     */
     public static function delete_assessment($params) {
         global $CFG, $DB;
         require_once($CFG->dirroot . '/blocks/evalcomix/classes/webservice_evalcomix_client.php');
         require_once($CFG->dirroot . '/blocks/evalcomix/classes/evalcomix_grades.php');
         $result = false;
         $where = (isset($params['where'])) ? $params['where'] : null;
+        $ws = (isset($params['ws'])) ? $params['ws'] : true;
         if (!empty($where) && is_array($where) && $assessments = $DB->get_records('block_evalcomix_assessments', $where)) {
             if ($result = $DB->delete_records('block_evalcomix_assessments', $where)) {
                 $courseid = (isset($params['courseid'])) ? $params['courseid'] : 0;
                 $cmid = (isset($params['cmid'])) ? $params['cmid'] : 0;
                 foreach ($assessments as $assessment) {
-                    $DB->delete_records('block_evalcomix_dr_pending', array('idassessment' => $assessment->idassessment));
-                    $DB->delete_records('block_evalcomix_dr_grade', array('idassessment' => $assessment->idassessment));
-                    block_evalcomix_webservice_client::delete_ws_assessment($assessment);
+                    $DB->delete_records('block_evalcomix_dr_pending', ['idassessment' => $assessment->idassessment]);
+                    $DB->delete_records('block_evalcomix_dr_grade', ['idassessment' => $assessment->idassessment]);
+                    if ($ws) {
+                        block_evalcomix_webservice_client::delete_ws_assessment($assessment);
+                    }
 
                     if (!empty($courseid) && !empty($cmid)) {
-                        if ($grade = $DB->get_record('block_evalcomix_grades', array('courseid' => $courseid,
-                            'cmid' => $cmid, 'userid' => $assessment->studentid))) {
-                            $finalgrade = block_evalcomix_grades::get_finalgrade_user_task(array('userid' => $assessment->studentid,
-                                'cmid' => $cmid, 'courseid' => $courseid));
+                        if (
+                            $grade = $DB->get_record('block_evalcomix_grades', ['courseid' => $courseid,
+                            'cmid' => $cmid, 'userid' => $assessment->studentid])
+                        ) {
+                            $finalgrade = block_evalcomix_grades::get_finalgrade_user_task(['userid' => $assessment->studentid,
+                                'cmid' => $cmid, 'courseid' => $courseid]);
                             if ($finalgrade !== null) {
-                                $DB->update_record('block_evalcomix_grades', array('id' => $grade->id, 'userid' => $grade->userid,
-                                    'cmid' => $grade->cmid, 'finalgrade' => $finalgrade, 'courseid' => $grade->courseid));
+                                $DB->update_record('block_evalcomix_grades', ['id' => $grade->id, 'userid' => $grade->userid,
+                                    'cmid' => $grade->cmid, 'finalgrade' => $finalgrade, 'courseid' => $grade->courseid]);
                             } else {
-                                $DB->delete_records('block_evalcomix_grades', array('id' => $grade->id));
+                                $DB->delete_records('block_evalcomix_grades', ['id' => $grade->id]);
                             }
                         }
                     }
@@ -279,5 +305,47 @@ class block_evalcomix_assessments extends block_evalcomix_object {
             }
         }
         return $result;
+    }
+
+    /**
+     * get_assessments_by_tasks
+     */
+    public static function get_assessments_by_tasks($taskids = [], $asessorids = []) {
+        global $DB;
+        $sql = "
+            SELECT a.*
+              FROM {block_evalcomix_assessments} a
+        ";
+
+        $where = [];
+        $params = [];
+
+        if (!empty($taskids)) {
+            [$insql, $inparams] = $DB->get_in_or_equal(
+                $taskids,
+                SQL_PARAMS_NAMED,
+                'task'
+            );
+
+            $where[] = "a.taskid $insql";
+            $params = array_merge($params, $inparams);
+        }
+
+        if (!empty($asessorids)) {
+            [$insql, $inparams] = $DB->get_in_or_equal(
+                $asessorids,
+                SQL_PARAMS_NAMED,
+                'assessor'
+            );
+
+            $where[] = "a.assessorid $insql";
+            $params = array_merge($params, $inparams);
+        }
+
+        if (!empty($where)) {
+            $sql .= ' WHERE ' . implode(' AND ', $where);
+        }
+
+        return $DB->get_records_sql($sql, $params);
     }
 }
